--- packages/elements/src/components/Breadcrumb/index.ts
@@ -1,1 +1,0 @@
-export { Breadcrumb, BreadcrumbItem, BreadcrumbSeparator } from '../../v2/src/components/breadcrumb'
