--- packages/elements/src/components/ImageCropper/index.tsx
@@ -1,1 +1,0 @@
-export * from '../../v2/src/components/image-cropper'
