+++ packages/elements/src/components/ImageCropperV2/__styles__/styles.ts
