+++ packages/elements/src/components/ImageCropperV2/__tests__/integration-helpers.test.ts
