+++ packages/elements/src/components/ImageCropperV2/integration-helpers.ts
