+++ packages/elements/src/components/ImageCropperV2/types.ts
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { FileInputProps } from '../../../../index'
+import { FileInputProps } from '../FileInput'
 
 export type CropBase = {
   width: number
