--- packages/elements/src/v2/src/components/index.ts
@@ -1,4 +1,0 @@
-export * from './typography'
-export * from './breadcrumb'
-export * from './image-cropper'
-export * from './modal'
