--- packages/elements/src/v2/src/components/typography/h1.test.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import { shallow } from 'enzyme'
-import { H1 } from './h1'
-
-describe('H1', () => {
-  it('should match snapshot', () => {
-    const wrapper = shallow(<H1 />)
-    expect(wrapper).toMatchSnapshot()
-  })
-})
