--- packages/elements/src/v2/src/components/typography/index.ts
@@ -1,1 +1,0 @@
-export { default as H1 } from './h1'
