--- packages/elements/src/v2/src/styles/index.ts
@@ -1,61 +1,0 @@
-export interface Theme {
-  h1: {
-    color: string
-  }
-}
-
-// Typography
-export const TYPOGRAPHY_TITLE_FONT_COLOR = 'black'
-export const TYPOGRAPHY_TITLE_FONT_WEIGHT = 600
-export const TYPOGRAPHY_TITLE_MARGIN_TOP = '1.2rem'
-export const TYPOGRAPHY_TITLE_MARGIN_BOTTOM = '0.5rem'
-
-// Colors
-export const WHITE = '#fff'
-export const BLACK = '#12263f'
-export const GREY_DARK = '#556986'
-export const GREY = '#95aac9'
-export const GREY_LIGHTER = '#eef3f9'
-export const GREY_LIGHTEST = '#f9fbfd'
-
-export const GREEN_LIGHTER = '#acf2bd'
-export const GREEN_LIGHTEST = '#e6ffed'
-export const RED_LIGHTER = '#fdb8c0'
-export const RED_LIGHTEST = '#ffeef0'
-
-export const REAPIT_DARK_BLUE = '#262f69'
-export const REAPIT_DARK_BLUE_DARKENED = '#1e2554'
-export const REAPIT_MID_BLUE = '#0061a8'
-export const REAPIT_LIGHT_BLUE = '#23a4de'
-export const REAPIT_LIGHTEST_BLUE = '#7bc9eb'
-export const REAPIT_LIGHTEST_BLUE_LIGHTENED = '#d4edf8'
-
-export const REAPIT_ORANGE = '#ec631b'
-export const REAPIT_LIME = '#cddb00'
-export const REAPIT_TEAL = '#006580'
-export const REAPIT_PLUMB = '#7a2c81'
-export const REAPIT_PURPLE = '#a4185c'
-export const REAPIT_GOLD = '#ffb71b'
-export const REAPIT_GREEN = '#a0c862'
-export const REAPIT_RED = '#d3033d'
-
-// Font
-export const FAMILY_SANS_SERIF = 'Roboto, Helvetica, Arial, sans-serif'
-export const FONT_SIZE = '14px'
-// Layout
-export const LAYOUT_BASE = '1rem'
-// Generic layout values that scale to rem base
-export const LAYOUT_QUARTER = 'calc(1rem / 4)'
-export const LAYOUT_THIRD = 'calc(1rem / 3)'
-export const LAYOUT_HALF = 'calc(1rem / 2)'
-export const LAYOUT_TWO_THIRD = 'calc(1rem * 0.67)'
-export const LAYOUT_THREE_QUARTER = 'calc(1rem * 0.75)'
-export const LAYOUT_ONE_QUARTER = 'calc(1rem * 1.25)'
-export const LAYOUT_ONE_HALF = 'calc(1rem * 1.5)'
-export const LAYOUT_DOUBLE = 'calc(1rem * 2)'
-export const LAYOUT_TWO_HALF = 'calc(1rem * 2.5)'
-export const LAYOUT_TRIPLE = 'calc(1rem * 3)'
-export const LAYOUT_QUADRUPLE = 'calc(1rem * 4)'
-export const LAYOUT_SEXTUPLE = 'calc(1rem * 6)'
-export const MENU_WIDTH = '18rem'
-export const PAGE_CONTAINER = 'calc(100vw - 18rem)'
